#43 NPE check for Elasticsearch homeDir in tests, creating Java objects for Kibana dashboards and their components.

+++ dcat-datastore/src/test/java/no/difi/dcat/datastore/DataStoreTest.java
@@ -108,7 +108,9 @@ public class DataStoreTest {
 			node.close();
 		}
 
-		FileUtils.forceDelete(homeDir);
+		if (homeDir != null) {
+			FileUtils.forceDelete(homeDir);
+		}
 
 		StoreConnection.reset();
